Add support for running the cli with provided repo or without
People (like me) might want to use the cli tool, but also be able to
adjust the hooks or do changes to the colorschemes and still run the cli
tool. This adds the option to provide a `--repo-dir` or `-d` option to
provide the path to `base16-shell`, otherwise it will use the files that
existed when the binary was compiled.